ztest: Separate unit test main into its own file
Zephyr headers tend to pull lots of other headers which
tend to create dependency loops and complicate treating
some parts of the code differently.
As the unit testing main() is actually a plain C main(),
meant to be compiled as a host binary, it needs some
special treatment compared to other Zephyr files.

So let's split it into its own file with minimal dependencies.
