+++ subsys/testsuite/ztest/src/ztest_unit_main.c
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2016 Intel Corporation
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <ztest_mock.h>
+#include <ztest_main_if.h>
+
+#ifndef KERNEL
+int main(void)
+{
+	z_init_mock();
+	test_main();
+	end_report();
+
+	return test_status;
+}
+#endif /* KERNEL */
